==> notebooks/5. engineer features for mapping (+remove probable duplicates).ipynb <==
## add custom data, generate tooltips & drop probable duplicates 

---

import os
import time

import geopandas
import numpy as np
import pandas as pd
from fuzzywuzzy import fuzz
from tqdm.notebook import tqdm
flatten = lambda l: [i for s in l for i in s]

%%capture
%%tqdm.pandas()

### read data

boot_locs = pd.read_csv('boot_locs.csv')
uni_locs = pd.read_csv('degs_gcoded.csv')
len(boot_locs), len(uni_locs)

# --> add more data here <--

boot_locs = pd.concat([boot_locs,
                       
                      # additional bootcamp data here <----
                       
                      ]).reset_index(drop=True)

uni_locs = pd.concat([uni_locs,
                      
                      # additional college data here <-----
                      
                      ]).reset_index(drop=True)

---

## Format features for tooltips (University data)

online_map = {'No': 'In Person', 'Yes': 'Online', 'Blended':'In Person + Online (blended)'}
uni_locs.online = uni_locs.online.map(online_map)

def parse_enroll(en):
    splt = str(en).split(' & ')
    if len(splt)==1: return en
    return ' & '.join(set(splt))
uni_locs.enrollment = uni_locs.enrollment.apply(parse_enroll).value_counts()

uni_locs.dropna(subset=['latitude'], inplace=True)

for row in uni_locs.index:
    
    dp = uni_locs.loc[row, 'dur_period']
    dn = uni_locs.loc[row, 'dur_number']
       
    try:
        dn = int(dn) # adjust months/yeasr
        if dn<=18:
            leng = f'{int(dn)}'
            dp = 'month'
        else:
            leng = f'{round(int(dn)/12, 1)}'.rstrip('.0')
            dp = 'year'

        uni_locs.loc[row, 'duration'] = f'{leng} {dp}s'            
    
    except:
        uni_locs.loc[row, 'duration'] = f'(?)'
        
    deg_spec = uni_locs.loc[row, 'degree'] # pretty degree
    deg_class = uni_locs.loc[row, 'class']
    commitment = uni_locs.loc[row, 'enrollment']
    commitment = (str(commitment)+' ') if 'nan'!=str(commitment).lower()!='(?)' else ''
    online = uni_locs.loc[row, 'online']  
    online = '| '+str(online) if 'nan'!=str(online).lower()!='(?)' else ''
    if str(deg_spec).lower()!='nan' and deg_spec[:5].lower() != deg_class[:5].lower():
        uni_locs.loc[row, 'pretty_degree'] = f'<b>{commitment}{deg_class} ({deg_spec})</b> {online}'
    else:
        uni_locs.loc[row, 'pretty_degree'] = f'<b>{commitment}{deg_class} </b> {online}'
    
uni_locs = uni_locs.fillna('(?)')

uni_locs.pretty_degree.value_counts().sample(10)

#### fix online indicator:

for ridx in uni_locs.index:
    
    if 'online' in uni_locs.loc[ridx, 'course_name'].lower():
        uni_locs.loc[ridx, 'online'] = 'Yes'
    

## add custom degrees data from African countries

pd.read_csv('africa_degs.csv',)[uni_locs.columns]

# I generated this data manually through research:
a_degs = pd.read_csv('african_masters_degs.csv',)[uni_locs.columns]

uni_locs = pd.concat([uni_locs, a_degs]).reset_index(drop=True)

a_boot = pd.read_csv('africa_boots.csv',)[boot_locs.columns]
boot_locs = pd.concat([boot_locs, a_boot]).reset_index(drop=True)


uni_locs.latitude = uni_locs.latitude.astype(float)
uni_locs.longitude = uni_locs.longitude.astype(float)

boot_locs.latitude = boot_locs.latitude.astype(float)
boot_locs.longitude = boot_locs.longitude.astype(float)

---

## Format features for tooltips (Bootcamp data)

boot_locs['pretty_camp'] = '<b>'+boot_locs['Commitment'].fillna('').apply(lambda x: x.replace(' ', '-' ).capitalize()
                                        )+ ' Bootcamp</b> | ' + boot_locs['Course Type'].fillna('')
boot_locs = boot_locs.fillna('(?)')

# drop bootcamps without one of these terms in the title

def topic_check(c):
    c = c.lower().replace('-', ' ')
    
    hot_words = ['data', 'stat', 'analy', 'busin', 'python'
                 'intell', 'informat', 'fintech', 'cybersecur',
                 'machine learn', 'machine-learn', 'optimiz',
                 'geosp', 'ter sci', 'guage proc']
    
    if any(kword in c for kword in hot_words):
        return True
    return False

for row_idx in boot_locs.index:
    c_name = boot_locs.loc[row_idx, 'course_name']
    if not topic_check(c_name): boot_locs.drop(row_idx, inplace=True)

---

## final duplicates check:

universities are the main problem

##### drop english language courses
uni_locs = uni_locs[~uni_locs.course_name.str.contains('English Language')]

# isolatae african progs (to save because manually added)
african_progs = uni_locs[uni_locs.institution=='African Institute for Mathematical Sciences']

len(uni_locs) + len(boot_locs)

# check for duplicates

verb = False # activate for tuning

def parse_title(title):
    title = str(title).split('(')[0].lower()
    return title

dropped = 0
scanned = 0
for ridx in uni_locs.index: # compare each uni program against all others
    scanned += 1 # to print progress 
    if ridx not in uni_locs.index: continue # already dropped
    
    # parameters to compare:
    rcourse = parse_title(uni_locs.loc[ridx, 'course_name']) # track name
    rinst = uni_locs.loc[ridx, 'institution']
    rclass = uni_locs.loc[ridx, 'class'] # bachelors, masters or phd
    blat = round(uni_locs.loc[ridx, 'latitude'], 3) # generalized location
    blng = round(uni_locs.loc[ridx, 'latitude'], 3)
    
    if 'African Institute for Mathematical Sciences' in rinst: continue
        # these were all manually added and should be preserved
    
    for cidx in uni_locs.drop(ridx).index:
        if ridx not in uni_locs.index: continue # (already dropped, don't check)
        
        # parameters to compare:
        ccourse = parse_title(uni_locs.loc[cidx, 'course_name'])
        cinst = uni_locs.loc[cidx, 'institution']
        cclass = uni_locs.loc[cidx, 'class']
        clat = round(uni_locs.loc[cidx, 'latitude'], 3) # same general location
        clng = round(uni_locs.loc[cidx, 'latitude'], 3)
        
        # combined match ratios:
        if fuzz.ratio(rinst, cinst) > 95 and rclass==cclass:
            if fuzz.partial_ratio(rcourse, ccourse) > 95:
                if blat==blng and clat==clng:
                    rnan = uni_locs.loc[ridx].isna().sum() + (uni_locs.loc[ridx]=='(?)').sum()
                    cnan = uni_locs.loc[cidx].isna().sum() + (uni_locs.loc[cidx]=='(?)').sum()
                    if rnan>cnan:
                        uni_locs.drop(ridx, inplace=True)
                        keeping = f'{cclass[:2]}: {ccourse} ({cinst}) - ({cnan} NA)'
                        dropping = f'{rclass[:2]}: {rcourse} ({rinst}) - ({rnan} NA)'
                    else:
                        uni_locs.drop(cidx, inplace=True)
                        dropping = f'{cclass[:2]}: {ccourse} ({cinst}) - ({cnan} NA)'
                        keeping = f'{rclass[:2]}: {rcourse} ({rinst}) - ({rnan} NA)'
                    dropped +=1
                    
                    if verb:
                        print('keeping:', keeping)
                        print('dropping:', dropping)
                        print('—'*80)
    
    if ridx%10==0 and not verb:
        print(f'\rDropped: {dropped} | Scanned: {scanned}/{len(uni_locs)+dropped}', end='')

# SAVE

uni_locs.to_csv('degs_to_map.csv', index=False)
boot_locs.to_csv('boot_to_map.csv', index=False)

len(uni_locs) + len(boot_locs)

uni_locs
