==> notebooks/7. overview statistics.ipynb <==
import json
import numpy as np
import pandas as pd

### read data

boot_locs = pd.read_csv('boot_to_map.csv')
uni_locs = pd.read_csv('degs_to_map.csv')
print(f'{len(boot_locs)} bootcamps & {len(uni_locs)} college/university degrees')
print(len(boot_locs)+len(uni_locs), 'total')

def get_country(g_info):
    try:
        g_json = json.loads(g_info.replace("'", "\""))
        return [s for s in g_json[0]['address_components'] if
                'country' in s['types']][0]['long_name']
    except: return g_info
    
uni_locs['country'] = uni_locs.gmaps_info.apply(get_country)

print(len(uni_locs), 'degrees across', uni_locs.country.nunique(), 'countries')

print('College/Uni Programs')
n_usa = uni_locs['country'].value_counts()[0]
print(f'  Programs in USA: {n_usa}\n  Outside USA: {len(uni_locs) - n_usa}')

top_cs = uni_locs['country'].value_counts()[:15].to_frame()
print('College/Uni Programs')
for i, country in enumerate(top_cs.index):
    print(f"  {i+1}. {country} ({top_cs.loc[country, 'country']})")

# ...
